DDP-5747 Move the composite answer component to a separate directory, make additional files for the template and styles
