+++ ddp-workspace/projects/ddp-sdk/src/lib/components/activityForm/answers/activity-composite-answer/activityCompositeAnswer.component.html
@@ -1,0 +1,28 @@
+<ddp-question-prompt [block]="block"></ddp-question-prompt>
+<ng-container *ngFor="let childBlockRow of childQuestionBlocks; let row = index">
+    <p *ngIf="row > 0" class="Title-text PageContent-subtitle Normal">
+        <span>{{ block.additionalItemText }}</span>
+        <button *ngIf="!readonly" mat-icon-button (click)="removeRow(row)">
+            <mat-icon class="ddp-close-button">close</mat-icon>
+        </button>
+    </p>
+    <div [ngClass]="['ddp-answer-container', setOrientationClass(block.childOrientation)]">
+        <ng-container *ngFor="let childBlock of childBlockRow; let column = index">
+            <ddp-activity-answer [ngClass]="[(childBlock.questionType == 'text'
+                                              && block.childOrientation !== 'VERTICAL') ? 'ddp-composite-text' : '',
+                                              'ddp-answer-field']"
+                                 [block]="childBlock"
+                                 [readonly]="readonly"
+                                 [validationRequested]="validationRequested"
+                                 (valueChanged)="updateValue(row, column, $event)"
+                                 (visibilityChanged)="updateVisibility($event)">
+        </ddp-activity-answer>
+        </ng-container>
+    </div>
+</ng-container>
+<button *ngIf="block.allowMultiple && !readonly"
+        type="button"
+        class="ButtonFilled ButtonFilled--white button button_medium button_secondary button_new-item"
+        (click)="addBlankRow()">
+    {{ block.addButtonText }}
+</button>
