+++ ddp-workspace/projects/ddp-sdk/src/lib/components/activityForm/answers/activity-composite-answer/activityCompositeAnswer.component.scss
@@ -1,0 +1,15 @@
+.Title-text {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+ }
+
+.vertical {
+  display: flex;
+  flex-direction: column;
+}
+
+.horizontal {
+  display: flex;
+  flex-direction: row;
+}
